% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{import}
\name{generics}
\alias{generics}
\alias{tidy}
\title{Tidy Generics}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{modelgenerics}{\code{\link[modelgenerics]{tidy}}}
}}

