% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapConnection-tidiers.R
\name{redcapConnection_tidiers}
\alias{redcapConnection_tidiers}
\alias{tidy.redcapApiConnection}
\title{Tidy a \code{redcapConnection} object}
\usage{
\method{tidy}{redcapApiConnection}(x, ...)
}
\arguments{
\item{x}{An object with class \code{redcapApiConnection} or 
\code{redcapDbConnection}.}

\item{...}{Arguments to pass to other methods.}
}
\description{
Makes a data frame containing the information used in a 
  REDCap API Connection.
}
